pyproject.toml
[tool.poetry]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = "A CLI tool to fetch PubMed research papers with non-academic authors"
authors = ["Your Name <your_email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.81"

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
